test: refactor fixtures file
